+++ androidApp/src/main/java/com/bupware/wedraw/android/ui/chatScreen/ChatScreen.kt
@@ -1,7 +1,6 @@
 package com.bupware.wedraw.android.ui.chatScreen
 
 import android.annotation.SuppressLint
-import android.util.Log
 import androidx.activity.compose.BackHandler
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.AnimatedVisibility
@@ -63,7 +62,7 @@ import com.bupware.wedraw.android.components.extra.DeviceConfig
 import com.bupware.wedraw.android.components.extra.GetDeviceConfig
 import com.bupware.wedraw.android.components.textfields.TextFieldMessage
 import com.bupware.wedraw.android.logic.models.Group
-import com.bupware.wedraw.android.logic.sessionData.sessionData
+import com.bupware.wedraw.android.logic.memoryData.sessionData
 import com.bupware.wedraw.android.theme.Lexend
 import com.bupware.wedraw.android.theme.blueVariant2WeDraw
 import com.bupware.wedraw.android.theme.redWeDraw
@@ -77,7 +76,7 @@ fun PreviewChatScreen(){
 
 
 @Composable
-fun ChatScreen(navController: NavController, groupId: Int, viewModel: ChatScreenViewModel = hiltViewModel()){
+fun ChatScreen(navController: NavController, groupId: Long, viewModel: ChatScreenViewModel = hiltViewModel()){
 
     LaunchedEffect(Unit){
         viewModel.groupId = groupId
